nordref v0.1.0 Nordref.Courses

The Courses context.

Functions

change_course(course)

Returns an %Ecto.Changeset{} for tracking course changes.

Examples

iex> change_course(course)
%Ecto.Changeset{source: %Course{}}

create_course(attrs \\ %{})

Creates a course.

Examples

iex> create_course(%{field: value})
{:ok, %Course{}}

iex> create_course(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_course(course)

Deletes a Course.

Examples

iex> delete_course(course)
{:ok, %Course{}}

iex> delete_course(course)
{:error, %Ecto.Changeset{}}

get_course!(id)

Gets a single course.

Raises Ecto.NoResultsError if the Course does not exist.

Examples

iex> get_course!(123)
%Course{}

iex> get_course!(456)
** (Ecto.NoResultsError)

list_courses()

Returns the list of courses.

Examples

iex> list_courses()
[%Course{}, ...]

release(course)

Release the given course.

This means, all previously reserved seats for the organizer of the course will be available for everyone else.

Examples

iex> release(course)
{:ok, %Course{}}

update_course(course, attrs)

Updates a course.

Examples

iex> update_course(course, %{field: new_value})
{:ok, %Course{}}

iex> update_course(course, %{field: bad_value})
{:error, %Ecto.Changeset{}}