nordref v0.1.0 Nordref.Registrations

The Registrations context.

Functions

create_registration(attrs \\ %{})

Creates a registration.

Examples

iex> create_registration(%{field: value})
{:ok, %Registration{}}

iex> create_registration(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_registration(registration)

Deletes a Registration.

Examples

iex> delete_registration(registration)
{:ok, %Registration{}}

iex> delete_registration(registration)
{:error, %Ecto.Changeset{}}

register(user, course)

Register the given user for the given course.

The user is allowed to participate in a course iff:

There is a seat available.

They have not been signed up for a course this season, except it is a G-course and the given course is the corresponding G-course.

Examples

iex> register(user, course)
{:ok, %Registration{}}

iex> register(user, course)
{:error, :not_allowed}

iex> register(user, course)
{:error, :not_available}

registrations_for_course(course)

Returns the list of registrations for the given course.

Examples

iex> get_registrations(course)
[%RegistrationView{}, ...]