nordref v0.1.0 NordrefWeb.Router.Helpers

Module with named helpers generated from NordrefWeb.Router.

Functions

club_path(conn_or_endpoint, action)

club_path(conn_or_endpoint, action, params)

club_path(conn_or_endpoint, action, id, params)

club_url(conn_or_endpoint, action)

club_url(conn_or_endpoint, action, params)

club_url(conn_or_endpoint, action, id, params)

course_path(conn_or_endpoint, action)

course_path(conn_or_endpoint, action, params)

course_path(conn_or_endpoint, action, id, params)

course_url(conn_or_endpoint, action)

course_url(conn_or_endpoint, action, params)

course_url(conn_or_endpoint, action, id, params)

license_path(conn_or_endpoint, action)

license_path(conn_or_endpoint, action, params)

license_path(conn_or_endpoint, action, id, params)

license_url(conn_or_endpoint, action)

license_url(conn_or_endpoint, action, params)

license_url(conn_or_endpoint, action, id, params)

path(data, path)

Generates the path information including any necessary prefix.

season_path(conn_or_endpoint, action)

season_path(conn_or_endpoint, action, params)

season_path(conn_or_endpoint, action, id, params)

season_url(conn_or_endpoint, action)

season_url(conn_or_endpoint, action, params)

season_url(conn_or_endpoint, action, id, params)

static_integrity(endpoint, path)

Generates an integrity hash to a static asset given its file path.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.

user_path(conn_or_endpoint, action)

user_path(conn_or_endpoint, action, params)

user_path(conn_or_endpoint, action, id, params)

user_url(conn_or_endpoint, action)

user_url(conn_or_endpoint, action, params)

user_url(conn_or_endpoint, action, id, params)